fix: address code review issues (dead code, duplication, panic safety)
- Remove unused pending_questions from GatewayState
- Extract shared handle_question_event() to reduce duplication
- Fix session_id[..8] panic risk with .min(8)
- Add sessionId fallback in WeCom question handler
- Add comment explaining WeCom question_id-as-key design

via [Happy](https://happy.engineering)
